Export translate function, that allows library users to use function outside a react component
