+++ src/index.js
@@ -2,5 +2,6 @@ import IntlReducer from './reducer'
 import withTranslate from './withTranslate'
 import IntlProvider from './provider'
 import * as IntlActions from './actions'
+import {translate as translateUtil} from "./utils"
 
-export { IntlReducer, withTranslate, IntlProvider, IntlActions }
+export { IntlReducer, withTranslate, IntlProvider, IntlActions, translateUtil }
